Locale support for cococo.
Preliminary support for locales (languages) for cococo.

Two new DB tables are added:

  * locales: contains infos about locales (code, name)
  * locales_strings: maps (key, locale ID) -> translated text

The new "id_locale" field of table "users" chooses the locale for a
specific user.

Regarding the back-end: LocaleManager is a new manager that makes it easy
to retrieve a translated string given a unique text key for the currently
logged in user. It also works "offline", given a key and a locale code.

Utils functions "T", "_T", "TE" and so on are shortcuts to use within the
front-end PHP pages. A new front-end configuration also determines what
locales are valid as for the front-end side.

Validation and signup/profile pages are up-to-date for this new feature.

Now we just need translators!